Fix bug causing crash after accidental response
When participants would accidentally hit 'enter/return' when inputting
their T1 or T2 response, the script would crash, because 'enter' is not
part of the string array of possible buttons, meaning it would be
indexed incorrectly.
Fixed this by de-activating 'enter' in PCL and only re-activate it after
a block has ended (after which participants should press enter to
resume the task).