Add option to terminate task on button press
When the "home" button is pressed, the experiment will terminate after
the current trial finishes. Because this task is self-paced (subjects
can take as long as they want to respond), it is hard to predict how
many trials will fit into x minutes of task performance. Now, the
experimenter can keep time and manually exit after x minutes have gone
by. This serves to equalize time on task for all participants and to
have the maximum amount trials for each participant.