Handle multiple data files per block
Sometimes the task crashed, after which it had to be restarted.
In these cases there are multiple data files for a single block
(numbered with 1, 2, etc.). For participants where this is the case,
the script now loads all the files and merges them, before extracting
outcome measures from the merged data frame.