Make trigger recoding robust to missing triggers
Sometimes EEG triggers can be missing, for example due to a conflict on
the parallel port (occurs pretty often at least once). Previously, in
the event of missing triggers, preprocessing code would crash when
trying to recode the T1 trigger to more meaningful values for analysis.
Now code is robust to missing triggers, and skips over recoding the T1
trigger if this is the case.