Save which preprocessing steps have been completed
When loading a partially pre-processed file, now checks which
preprocessing steps were done to make that file, and saves this
information in the new file. That way, it is always possible to
reconstruct the entire pipeline in the final preprocessed file, even if
that file was based on intermediate loads and saves.