Select segment of data for pre-processing
Adds step in the pipeline right after loading the data, to select only
a portion of the continuous data in the tDCS block. Specifically, data
before end of the ramp-down and after end of the ramp-up are discarded,
to avoid filtering artifacts from the DC offsets that could be there
(especially after ramp-down).