Force all event markers to be strings
This converts all event markers to strings in preproc_recode triggers,
and changes all relevant functions to handle strings instead of
doubles.

EEGLAB can insert 'boundary' events, for example when you select a
portion of the continuous data. This is done in preproc_segment_data
for the tDCS blocks, to cut out the ramp-up/down periods. While event
markers are normally doubles, if there is a boundary event EEGLAB
will convert all other events to strings as well, which the code
was not equipped to deal with.