Analysis of T1 and T2 swaps
Modify MATLAB behavioral analysis script of the pilot data, to also
plot proportion of trials on which T1 and T2 were swapped, actual T2
was reported as T1, and actual T1 was reported as T2. Especially
in the latter case it can be argued that these are also blink(-like)
trials.