Modify T1 triggers occurring during ramp-down
Unknown whether data during ramp-down can be used for analysis, as it
might contain (more) artifacts. Would be a shame to discard a minute of
data though, so change the markers of T1s occurring in the ramp-down
period, so it can be decided later whether to include them or not.