Make preprocessing compatible with newest eeglab
Choice of running with either version 8 (confirmed working on MATLAB
r2010b) or version 13_5_4b (confirmed working on MATLAB
r2012b and r2015a). Which version is to be used should now be specified
by the user.

Main difference between the two: Importing the data works correctly in
8; in the newer version a large number is added to each event value,
which has to be corrected for.