Use new filter from firfilt plugin
The old filter ('pop_eegfilt', eeglab version 8) is deprecated -
defaults were poor and could cause excessive filter ringing. It is
also incredibly slow for high-pass filtering in particular, because
the transition band width was not automatically determined. So if it
exists on the path, 'pop_eegfiltnew' will be used instead.