Throw warning, not error when no trials to reject
Previously, function intentionally crashed when it could not find list
of trials to reject. However, sometimes it might not be necessary to
reject any trials. So issue a warning instead, to alert the user to
double check whether indeed no trials should be rejected (or the trial
rejection step was inadvertently skipped).