Blind preprocessing pipeline (first half of data!)
To allow for blind data analysis, everything is now defined only in
terms of tDCS codes (settings 'B' and 'D' on the device), no longer in
terms of stimulation type ('anodal/cathodal', each coded by one of the
letters).

N.B. This coding scheme is only valid for data from S01 through S21!
Afterwards the device was reprogrammed, and files are coded with 'D'
(which does not necessarily correspond to the same stimulation type as
'D' in the first half!) and 'I'.