Interpolate epochs before channels
Interpolation of channels on single epochs should come before
interpolation of channels in their entirety (order before was
reversed).

The pipeline was actually setup this way already:
channels to be interpolated entirely are excluded from single-epoch
interpolation, so that bad data is not included in the interpolation.
With the incorrect order, noisy single epochs (that will be
interpolated) are actually included in the channel interpolation, which
makes for a (slightly) messier result.