Fix error in T2|T1 percentage accuracy
In all results so far, the wrong denominator for T2|T1 ("T2 given T1")
accuracy was used. Instead of the trials where T1 was
reported correctly, all trials were used to compute % correct. So
instead of T2|T1, this was simply the percentage of trials in which T2
and T1 were both reported correctly (out of all trials).

Now fixed to true T2|T1 accuracy, meaning the percentage of
trials in which T2 and T1 were reported correctly, out of the subset
of trials in which T1 was reported correctly.

Also added a T2 accuracy column (% of trials where T2 was reported
correctly, out of all trials) in case that might be useful in the
future.